# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bidly API",
    "version": "1.0.0",
    "description": "Generated from Bidly Zod command schemas. No manually duplicated contract."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/v1/me": {
      "get": {
        "operationId": "get__api_v1_me",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "user_id",
                    "roles"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories": {
      "get": {
        "operationId": "get__api_v1_categories",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "version_id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "market_type": {
                            "type": "string",
                            "enum": [
                              "SWITCH",
                              "CAPACITY",
                              "BULK",
                              "LEAD"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "version_id",
                          "slug",
                          "market_type"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          }
        ]
      }
    },
    "/api/v1/offers/{id}/accept": {
      "post": {
        "operationId": "post__api_v1_offers_id_accept",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "offer_id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "reservation_id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "status": {
                      "type": "string",
                      "const": "ACCEPTED"
                    }
                  },
                  "required": [
                    "offer_id",
                    "reservation_id",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 8,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "request_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "offerId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 128
                  }
                },
                "required": [
                  "offerId",
                  "expectedVersion",
                  "idempotencyKey"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 16,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9._:-]+$"
            }
          }
        ]
      }
    }
  }
}